+++ src/functions/arrays/isArray.ts
@@ -1,0 +1,3 @@
+import type { isArray_F } from './types';
+
+export const isArray: isArray_F = value => Array.isArray(value);