+++ src/functions/arrays/types.ts
@@ -1,0 +1,3 @@
+export type isArray_F = <T>(value: unknown) => value is T[];
+
+export type ToArray_F = <T>(obj?: any) => T[];