+++ src/functions/common/toArray.ts
@@ -1,0 +1,5 @@
+import type { ObjectToArray_F } from './types';
+
+export const objectToArray: ObjectToArray_F = obj => {
+  return Object.values(obj);
+};