+++ src/functions/index.ts
@@ -1,4 +1,6 @@
+export * from './arrays';
 export * from './common';
 export * from './functions';
 export * from './numbers';
+export * from './strings';
 export * from './types';