+++ src/functions/types/isDefined.build.test.ts
@@ -1,15 +1,24 @@
-import { createTests } from '@bemedev/vitest-extended';
+import { this1 } from '@bemedev/build-tests/constants';
 
-import { isDefined } from 'this-gen-1';
+import { t } from '@bemedev/types';
+import { isDefined as func } from './isDefined';
 
 describe('isDefined', () => {
-  const useTests = createTests(isDefined);
+  let isDefined = t.anify<typeof func>();
 
-  useTests(
-    { expected: false, invite: 'no Value' },
-    { expected: false, invite: 'undefined', parameters: [undefined] },
-    { expected: true, invite: 'number : 9', parameters: 9 },
-    { expected: false, invite: 'boolean : false', parameters: false },
-    { expected: true, invite: 'boolean : true', parameters: [true] },
-  );
+  beforeAll(() => {
+    return import(this1).then(module => {
+      isDefined = module.isDefined;
+    });
+  });
+
+  test.each([
+    ['No value', undefined, false],
+    ['string', 'str', true],
+    ['number', 50, true],
+    ['boolean', true, true],
+    ['boolean', false, false],
+  ] as const)('#%# => %s', (_, args, expected) => {
+    expect(isDefined(args)).toBe(expected);
+  });
 });