+++ CHANGE_LOG.md
@@ -3,6 +3,19 @@
 <br/>
 <br/>
 
+## [0.0.2] 2025/01/21
+
+### Added
+
+- Added utility functions for string manipulation.
+
+### Fixed
+
+- Export all functions
+
+<br/>
+<br/>
+
 ## [0.0.1] 2025/01/11
 
 <br/>